requirements.txt
Django==1.7.2
django-braces==1.4.0
django-guardian==1.2.5
django-markdown-deux==1.0.5
django-ratings==0.3.7
django-widget-tweaks==1.3
markdown2==2.3.0
Pillow==2.7.0
pysqlite==2.6.3
pytz==2014.10
six==1.9.0
yolk==0.4.3
